feat: get allowed routes on headless controller
