+++ apps/storefront/src/components/layout/B3RenderRouter.tsx
@@ -4,12 +4,8 @@ import { Outlet, Route, Routes, useLocation, useNavigate } from 'react-router-do
 import { RegisteredProvider } from '@/pages/Registered/context/RegisteredContext';
 import { type SetOpenPage } from '@/pages/SetOpenPage';
 import { GlobalContext } from '@/shared/global';
-import {
-  firstLevelRouting,
-  getAllowedRoutes,
-  RouteFirstLevelItem,
-  RouteItem,
-} from '@/shared/routes';
+import { getAllowedRoutes, RouteFirstLevelItem, RouteItem } from '@/shared/routeList';
+import { firstLevelRouting } from '@/shared/routes';
 import { getPageTranslations, useAppDispatch } from '@/store';
 import { channelId } from '@/utils';
