+++ apps/storefront/src/shared/dynamicallyVariable/context/index.ts
@@ -2,12 +2,12 @@ import { createContext } from 'react';
 
 import {
   DynamicallyVariableAction,
-  DynamicallyVariableContext,
+  DynamicallyVariableContextInterface,
   DynamicallyVariableState,
   initState,
 } from './config';
 
-export const DynamicallyVariableedContext = createContext<DynamicallyVariableContext>({
+export const DynamicallyVariableContext = createContext<DynamicallyVariableContextInterface>({
   state: initState,
   dispatch: () => {},
 });
