+++ apps/storefront/src/shared/dynamicallyVariable/index.ts
@@ -1,2 +1,2 @@
-export { DynamicallyVariableedContext } from './context';
+export { DynamicallyVariableContext } from './context';
 export { default as DynamicallyVariableProvider } from './context/Provider';
